# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Flipeet",
      "description": "Flipeet Books E-Commerce Store Apis Documentation",
      "version": "1.0.0"
    },
    "schemes": [
      "http"
    ],
    "basePath": "/",
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/user/register": {
        "post": {
          "summary": "Register a new User itself",
          "description": "When a cutomer visit our site and want to buy or sale a book than he must Register itSelf in this site with personal information",
          "parameters":[
            {
              "name": "firstName",
              "in": "query",
              "description": "First Name of User.",
              "type": "String"

            },
            {
              "name": "lastName",
              "in": "query",
              "description": "Last Name of User.",
              "type": "String"

            },
            {
              "name": "email",
              "in": "query",
              "description": "Email of User.",
              "type": "String"

            },
            {
              "name": "password",
              "in": "query",
              "description": "Password of User.",
              "type": "String"

            },
            {
              "name": "conform password",
              "in": "query",
              "description": "Confirm Password of User",
              "type": "String"

            },
            {
              "name": "role",
              "enum": ["admin", "user"],
              "in": "query",
              "description": "there are two roles of user first is 'user' and Second is 'admin'",
              "type": "String"

            },
            {
              "name": "profile_image",
              "consumes": "multipart/form-data",
              "in": "query",
              "description": "Image of User",
              "type": "String"

            },
            {
              "name": "accountType",
              "enum": ["student", "non-student"],
              "in": "query",
              "description": "There two accountType first is student, and second is non-student .",
              "type": "String"

            },
            {
              "name": "school",
              "in": "query",
              "description": "if the user id student then this field is used to data.",
              "type": "String"

            },
            {
              "name": "city",
              "in": "query",
              "description": "City of User.",
              "type": "String"

            },
            {
              "name": "tickets",
              "in": "query",
              "description": "Every user give by default 5 tickets.",
              "type": "Number"

            },
            {
              "name": "status",
              "in": "query",
              "description": "There three status of user but in there endpoint the default status is deactive.",
              "type": "String"

            }
            

          ],
          "tags": [
            "User"
          ],

          "responses": {
            "200": {
              "description": "Personal Information of User"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/verify/:_id": {
        "get": {
          "summary": "Verify the user account.",
          "description": "This endpoint use to verify the user account when verification email send to user than is endponit is used.",
          "parameters": [
            {
              "name":"_id",
              "description": "This is user to identify the user.",
              "type":"String",
              "in":"query"

            }
          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User Account Verify Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/forgetPassword": {
        "post": {
          "summary": "This endpoint is used to forget user password.",
          "description": "If the user don't remembered the password than User put email and verfication token send to user.",
          "parameters": [
            {
              "name":"email",
              "description": "This is used to input from their email from user and send email.",
              "type":"String"

            }
          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "Reset Token Send to User Email"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/resetPassword/:token": {
        "put": {
          "summary": "Create new password of user.",
          "description": "If user has correct reset token then user create user password.",
          "parameters": [
            {
              "name": "token",
              "description": "This is reset token that is send from server forget password api and verify this token.",
              "type": "String",
              "in":"query"

            },
            {
              "name": "password",
              "description": "User Password.",
              "type": "String"

            },
            {
              "name": "conform password",
              "description": "User confirm Password",
              "type": "String"
            }

        ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "Reset User Password Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/active-account": {
        "post": {
          "summary": "Activate the user status.",
          "description": "If the user status is 'deactive or suspend than this api is used to send email.",
          "parameters": [
            {
              "name": "email",
              "description": "User email used to send Email.",
              "type": "String"

            }

        ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User Email Send Successfully."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/login": {
        "post": {
          "summary": "Login user.",
          "description": "When the user register account sussfully then he able to login.",
          "parameters": [
            {
              "name": "email",
              "description": "User email used to verify the user account.",
              "type": "String"

            },
            {
              "name": "password",
              "description": "User password is used to matach password with email..",
              "type": "String"

            }

          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "if the user input data is correct than user login successfully loggedin."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/user/edit": {
        "put": {
          "summary": "Update User Profile",
          "description": "When the user register account sussfully then he able to login.",
          "parameters": [
            {
              "name": "firstName",
              "description": "First Name of User.",
              "type": "String"

            }, {
              "name": "lastName",
              "description": "Last Name of User.",
              "type": "String"

            }, {
              "name": "email",
              "description": "Email of User.",
              "type": "String"

            }, {
              "name": "password",
              "description": "Password of User.",
              "type": "String"

            }, {
              "role": "conform password",
              "description": "Confirm Password of User",
              "type": "String"

            }, {
              "name": "role",
              "in": "query",
              "enum": ["admin", "user"],
              "description": "there are two roles of user first is 'user' and Second is 'admin'",
              "type": "String"

            }, {
              "name": "profile_image",
              "consumes": "multipart/form-data",
              "in": "query",
              "description": "Image of User",
              "type": "String"

            }, {
              "name": "accountType",
              "enum": ["student", "non-student"],
              "description": "There two accountType first is student, and second is non-student .",
              "type": "String"

            }, {
              "name": "school",
              "in": "query",
              "description": "if the user id student then this field is used to data.",
              "type": "String"

            }, {
              "name": "city",
              "in": "query",
              "description": "City of User.",
              "type": "String"

            }, {
              "name": "tickets",
              "in": "query",
              "description": "Every user give by default 5 tickets.",
              "type": "Number"

            }, {
              "name": "status",
              "in": "query",
              "description": "There three status of user but in there endpoint the default status is deactive.",
              "type": "String"

            }

          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User profile update"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/userTickets": {
        "get": {
          "summary": "Get Ticket Details of User.",
          "description": "This end point is used to get details of tickets to admin side.",
          "parameters": [{
              "name": "_id",
              "description": "Authenticated user id.",
              "type": "String"

            }

          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User Tickets Details"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/user": {
        "get": {
          "summary": "Send Login User data.",
          "description": "This endpoint is used to send data of logged in user.",
          "parameters": [],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "Logged in user data."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/suspendUser/:id": {
        "put": {
          "summary": "Suspend user Account.",
          "description": "This endpoint is used to change user status.",
          "parameters": [
            {
              "name": "id",
              "description": "Authenticated user id used to suspend the status of logged in user.",
              "type": "String",
              "in":"query"

            }
          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User Account suspended."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/user/delete/:id": {
        "delete": {
          "summary": "Delete User Account.",
          "description": "This endpoint is used to delete the user account.",
          "parameters": [{
            "name": "id",
            "description": "Authenticated user id used to delete user.",
            "type": "String",
            "in": "query"

          }],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "User Account Deleted."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/editSiteUser/:id": {
        "put": {
          "summary": "Edit user by admin.",
          "description": "This endpoint is used that admin can edit user details.",
          "parameters": [
            {
            "name": "id",
            "description": "user id used to delete user.",
            "type": "String",
            "in": "query"
            },
            {
              "name": "firstName",
              "description": "First Name of User.",
              "type": "String"

            }, {
              "name": "lastName",
              "description": "Last Name of User.",
              "type": "String"

            }, {
              "name": "email",
              "description": "Email of User.",
              "type": "String"

            }, {
              "name": "password",
              "description": "Password of User.",
              "type": "String"

            }, {
              "role": "conform password",
              "description": "Confirm Password of User",
              "type": "String"

            }, {
              "name": "role",
              "in": "query",
              "enum": ["admin", "user"],
              "description": "there are two roles of user first is 'user' and Second is 'admin'",
              "type": "String"

            }, {
              "name": "profile_image",
              "consumes": "multipart/form-data",
              "in": "query",
              "description": "Image of User",
              "type": "String"

            }, {
              "name": "accountType",
              "enum": ["student", "non-student"],
              "description": "There two accountType first is student, and second is non-student .",
              "type": "String"

            }, {
              "name": "school",
              "in": "query",
              "description": "if the user id student then this field is used to data.",
              "type": "String"

            }, {
              "name": "city",
              "in": "query",
              "description": "City of User.",
              "type": "String"

            }, {
              "name": "tickets",
              "in": "query",
              "description": "Every user give by default 5 tickets.",
              "type": "Number"

            }, {
              "name": "status",
              "in": "query",
              "description": "There three status of user but in there endpoint the default status is deactive.",
              "type": "String"
            }

          ],
          "tags": [
            "User"
          ],
          "responses": {
            "200": {
              "description": "Admin Edit User Account."
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
    }
  }